+++ Sources/KeyboardKit/Gestures/SpaceCursorDragGestureHandler.swift
@@ -15,26 +15,37 @@ import CoreGraphics
  */
 open class SpaceCursorDragGestureHandler: DragGestureHandler {
 
-    /**
-     Create a handler space cursor drag gesture handler.
-
-     - Parameters:
-       - context: The keyboard context to use.
-       - feedbackHandler: The feedback handler to use.
-       - sensitivity: The drag sensitivity.
-     */
+    @available(*, deprecated, message: "Use the keyboardContext initializer instead.")
     public init(
         context: KeyboardContext,
         feedbackHandler: KeyboardFeedbackHandler,
-        sensitivity: SpaceDragSensitivity = .medium) {
-        self.context = context
+        sensitivity: SpaceDragSensitivity = .medium
+    ) {
+        self.keyboardContext = context
         self.feedbackHandler = feedbackHandler
         self.sensitivity = sensitivity
     }
 
-    public let context: KeyboardContext
-    public let feedbackHandler: KeyboardFeedbackHandler
+    /**
+     Create a handler space cursor drag gesture handler.
 
+     - Parameters:
+       - keyboardContext: The keyboard context to use.
+       - feedbackHandler: The feedback handler to use.
+       - sensitivity: The drag sensitivity, by default ``SpaceDragSensitivity/medium``.
+     */
+    public init(
+        keyboardContext: KeyboardContext,
+        feedbackHandler: KeyboardFeedbackHandler,
+        sensitivity: SpaceDragSensitivity = .medium
+    ) {
+        self.keyboardContext = keyboardContext
+        self.feedbackHandler = feedbackHandler
+        self.sensitivity = sensitivity
+    }
+
+    public let keyboardContext: KeyboardContext
+    public let feedbackHandler: KeyboardFeedbackHandler
     public let sensitivity: SpaceDragSensitivity
 
     public var currentDragStartLocation: CGPoint?
@@ -44,20 +55,26 @@ open class SpaceCursorDragGestureHandler: DragGestureHandler {
      Handle a drag gesture on space, which by default should
      move the cursor left and right after a long press.
      */
-    public func handleDragGesture(from startLocation: CGPoint, to currentLocation: CGPoint) {
+    public func handleDragGesture(
+        from startLocation: CGPoint,
+        to currentLocation: CGPoint
+    ) {
         tryStartNewDragGesture(from: startLocation, to: currentLocation)
         let dragDelta = startLocation.x - currentLocation.x
         let textPositionOffset = Int(dragDelta / CGFloat(sensitivity.points))
         guard textPositionOffset != currentDragTextPositionOffset else { return }
         let offsetDelta = textPositionOffset - currentDragTextPositionOffset
-        context.textDocumentProxy.adjustTextPosition(byCharacterOffset: -offsetDelta)
+        keyboardContext.textDocumentProxy.adjustTextPosition(byCharacterOffset: -offsetDelta)
         currentDragTextPositionOffset = textPositionOffset
     }
 }
 
 private extension SpaceCursorDragGestureHandler {
 
-    func tryStartNewDragGesture(from startLocation: CGPoint, to currentLocation: CGPoint) {
+    func tryStartNewDragGesture(
+        from startLocation: CGPoint,
+        to currentLocation: CGPoint
+    ) {
         let isNewDrag = currentDragStartLocation != startLocation
         currentDragStartLocation = startLocation
         guard isNewDrag else { return }