+++ Sources/KeyboardKit/_Deprecated/Gestures+Deprecated.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+public extension SpaceCursorDragGestureHandler {
+
+    @available(*, deprecated, renamed: "keyboardContext")
+    var context: KeyboardContext { keyboardContext }
+}